Rename to bikeRentalServiceDirectory
