+++ docs/sandbox/BikeRentalServiceDirectory.md
@@ -1,0 +1,14 @@
+# Bike Rental Service Directory API support.
+
+## Contact Info
+- Gard Mellemstrand, Entur, Norway
+
+## Changelog
+- Initial implementation of bike share updater API support
+
+## Documentation
+
+This adds support for the GBFS service directory endpoint component located at https://github.com/entur/bikeservice
+ 
+### Configuration
+To enable this you need to specify a url for the `bikeRentalServiceDirectory` in the `router-config.json`
