--- docs/sandbox/BikeShareUpdaterEndpoint.md
@@ -1,14 +1,0 @@
-# Bike share updater API support.
-
-## Contact Info
-- Gard Mellemstrand, Entur, Norway
-
-## Changelog
-- Initial implementation of bike share updater API support
-
-## Documentation
-
-This adds support for the GBFS endpoint component located at https://github.com/entur/bikeservice
- 
-### Configuration
-To enable this you need to specify a url for the `bikeShareUpdaterEndpoint` in the `router-config.json`
