+++ src/ext/java/org/opentripplanner/ext/bikerentalservicedirectory/BikeRentalServiceDirectoryFetcher.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.updaterendpointfetcher;
+package org.opentripplanner.ext.bikerentalservicedirectory;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -17,9 +17,9 @@ import java.util.List;
  * Fetches GBFS endpoints from the Bikeservice component located at
  * https://github.com/entur/bikeservice, which is an API for aggregating GBFS endpoints.
  */
-public class GBFSUpdaterEndpointFetcher {
+public class BikeRentalServiceDirectoryFetcher {
 
-  private static final Logger LOG = LoggerFactory.getLogger(GBFSUpdaterEndpointFetcher.class);
+  private static final Logger LOG = LoggerFactory.getLogger(BikeRentalServiceDirectoryFetcher.class);
 
   private static final int DEFAULT_FREQUENCY_SEC = 15;
